Restructured code for future development
