+++ client/index.html
@@ -15,17 +15,17 @@
 	
 	<script src="//cdn.socket.io/socket.io-1.3.7.js"></script>
 	
-	<script src="/telephone.module.js"></script>
-	<script src="/game.service.js"></script>
-	<script src="/drawing.directive.js"></script>
-	<script src="/waiting.controller.js"></script>
-	<script src="/waiting.directive.js"></script>
-	<script src="/login.controller.js"></script>
-	<script src="/login.directive.js"></script>
-	<script src="/guess.controller.js"></script>
-	<script src="/guess.directive.js"></script>
-	<script src="/draw.controller.js"></script>
-	<script src="/draw.directive.js"></script>
+	<script src="/client/js/telephone.module.js"></script>
+	<script src="/client/js/game.service.js"></script>
+	<script src="/client/js/drawing.directive.js"></script>
+	<script src="/client/js/waiting.controller.js"></script>
+	<script src="/client/js/waiting.directive.js"></script>
+	<script src="/client/js/login.controller.js"></script>
+	<script src="/client/js/login.directive.js"></script>
+	<script src="/client/js/guess.controller.js"></script>
+	<script src="/client/js/guess.directive.js"></script>
+	<script src="/client/js/draw.controller.js"></script>
+	<script src="/client/js/draw.directive.js"></script>
   </head>
   <body>
 	<login></login>
