+++ client/js/drawing.directive.js
