+++ client/js/waiting.controller.js
