+++ client/templates/draw.tpl.html
