+++ client/templates/guess.tpl.html
