+++ client/templates/waiting.tpl.html
